requirements.txt
# --- LLM Framework ---
langchain==1.2.0
langchain-core==1.2.5
langchain-community==0.4.1
langchain-openai==1.1.6

# --- Model & Tokenizer ---
openai==2.14.0
tiktoken==0.12.0

# --- Web Framework & API ---
fastapi==0.128.0
uvicorn==0.40.0
jinja2==3.1.6
python-multipart==0.0.21
aiofiles==25.1.0

# --- Vector Database & Data Processing ---
faiss-cpu==1.13.2
pypdf==6.5.0
python-dotenv==1.2.1

# --- More ---
-e .